DF-8660 review fixes
Former-commit-id: 3aee780aa629925f193afc584eb3e624f3770010

+++ Sources/Internal/Render/Highlevel/Landscape.cpp
@@ -175,7 +175,7 @@ int16 Landscape::AllocateQuadVertexBuffer(LandscapeQuad* quad)
         }
     }
 
-    uint32 vBufferSize = verticesCount * sizeof(LandscapeVertex);
+    uint32 vBufferSize = static_cast<uint32>(verticesCount * sizeof(LandscapeVertex));
 
     rhi::VertexBuffer::Descriptor desc;
     desc.size = vBufferSize;
